# docs/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiEnvelope_SearchData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[SearchData]",
        "type": "object"
      },
      "ApiEnvelope_SearchInspectData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchInspectData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[SearchInspectData]",
        "type": "object"
      },
...
      "ApiEnvelope_TargetPostsData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetPostsData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[TargetPostsData]",
        "type": "object"
      },
      "ApiEnvelope_TargetSocialTimelineData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetSocialTimelineData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[TargetSocialTimelineData]",
        "type": "object"
      },
      "ApiEnvelope_TokenCaseData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenCaseData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[TokenCaseData]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchData": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/SearchPageData"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "target_candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Target Candidates",
            "type": "array"
          }
        },
        "required": [
          "query",
          "page",
          "target_candidates",
          "items"
        ],
        "title": "SearchData",
        "type": "object"
      },
      "SearchInspectAmbiguousData": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Candidates",
            "type": "array"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/SearchInspectTopicSummaryData"
          }
        },
        "required": [
          "candidates",
          "summary",
          "items"
        ],
        "title": "SearchInspectAmbiguousData",
        "type": "object"
      },
      "SearchInspectData": {
        "additionalProperties": false,
        "properties": {
          "ambiguous_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchInspectAmbiguousData"
              },
              {
                "type": "null"
              }
            ]
          },
          "query": {
            "$ref": "#/components/schemas/SearchInspectQueryData"
          },
          "resolver": {
            "$ref": "#/components/schemas/SearchInspectResolverData"
          },
          "token_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenCaseData"
              },
              {
                "type": "null"
              }
            ]
          },
          "topic_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchInspectTopicData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "query",
          "resolver",
          "token_result",
          "topic_result",
          "ambiguous_result"
        ],
        "title": "SearchInspectData",
        "type": "object"
      },
      "SearchInspectQueryData": {
        "additionalProperties": false,
        "properties": {
          "normalized_q": {
            "title": "Normalized Q",
            "type": "string"
          },
          "q": {
            "title": "Q",
            "type": "string"
          },
          "result_kind": {
            "enum": [
              "token_result",
              "topic_result",
              "ambiguous_result",
              "empty_result"
            ],
            "title": "Result Kind",
            "type": "string"
          },
          "window": {
            "title": "Window",
            "type": "string"
          }
        },
        "required": [
          "q",
          "normalized_q",
          "window",
          "result_kind"
        ],
        "title": "SearchInspectQueryData",
        "type": "object"
      },
      "SearchInspectResolverData": {
        "additionalProperties": false,
        "properties": {
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "selected_target": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Target"
          },
          "target_candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Target Candidates",
            "type": "array"
          }
        },
        "required": [
          "target_candidates",
          "selected_target",
          "reasons"
        ],
        "title": "SearchInspectResolverData",
        "type": "object"
      },
...
      "SearchPageData": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "returned_count": {
            "title": "Returned Count",
            "type": "integer"
          }
        },
        "required": [
          "returned_count",
          "has_more",
          "next_cursor"
        ],
        "title": "SearchPageData",
        "type": "object"
      },
...
      "TargetPostsData": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "query": {
            "$ref": "#/components/schemas/TargetPostsQueryData"
          },
          "returned_count": {
            "title": "Returned Count",
            "type": "integer"
          },
          "score_window": {
            "$ref": "#/components/schemas/TargetPostsScoreWindowData"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "query",
          "score_window",
          "total_count",
          "returned_count",
          "has_more",
          "next_cursor",
          "items"
        ],
        "title": "TargetPostsData",
        "type": "object"
      },
...
      "TargetSocialTimelineData": {
        "additionalProperties": false,
        "properties": {
          "authors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Authors",
            "type": "array"
          },
          "buckets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Buckets",
            "type": "array"
          },
          "cascade": {
            "additionalProperties": true,
            "title": "Cascade",
            "type": "object"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "market_candles": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Candles"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "posts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Posts",
            "type": "array"
          },
          "query": {
            "$ref": "#/components/schemas/TargetSocialTimelineQueryData"
          },
          "returned_count": {
            "title": "Returned Count",
            "type": "integer"
          },
          "stages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Stages",
            "type": "array"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "query",
          "summary",
          "market_candles",
          "stages",
          "buckets",
          "authors",
          "posts",
          "cascade",
          "returned_count",
          "has_more",
          "next_cursor"
        ],
        "title": "TargetSocialTimelineData",
        "type": "object"
      },
...
      "TokenCaseData": {
        "additionalProperties": false,
        "properties": {
          "current_radar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenRadarFactRowData"
              },
              {
                "type": "null"
              }
            ]
          },
          "market_live": {
            "additionalProperties": true,
            "title": "Market Live",
            "type": "object"
          },
          "posts": {
            "additionalProperties": true,
            "title": "Posts",
            "type": "object"
          },
          "profile": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "target": {
            "additionalProperties": true,
            "title": "Target",
            "type": "object"
          },
          "timeline": {
            "additionalProperties": true,
            "title": "Timeline",
            "type": "object"
          }
        },
        "required": [
          "target",
          "profile",
          "timeline",
          "posts",
          "market_live",
          "current_radar"
        ],
        "title": "TokenCaseData",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Tracefold",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/search": {
      "get": {
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "",
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "24h",
              "title": "Window",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SearchData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "api"
        ]
      }
    },
    "/api/search/inspect": {
      "get": {
        "operationId": "search_inspect_api_search_inspect_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "24h",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SearchInspectData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Inspect",
        "tags": [
          "api"
        ]
      }
    },
...
    "/api/target-posts": {
      "get": {
        "operationId": "target_posts_api_target_posts_get",
        "parameters": [
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "5m",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "default": "current_window",
              "title": "Range",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "",
              "title": "Cursor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_TargetPostsData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Target Posts",
        "tags": [
          "api"
        ]
      }
    },
    "/api/target-social-timeline": {
      "get": {
        "operationId": "target_social_timeline_api_target_social_timeline_get",
        "parameters": [
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "1h",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "",
              "title": "Cursor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_TargetSocialTimelineData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Target Social Timeline",
        "tags": [
          "api"
        ]
      }
    },
    "/api/token-case": {
      "get": {
        "operationId": "token_case_api_token_case_get",
        "parameters": [
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "1h",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "posts_limit",
            "required": false,
            "schema": {
              "default": 24,
              "title": "Posts Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_TokenCaseData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Token Case",
        "tags": [
          "api"
        ]
      }
    },
...
  }
}